requirements.txt
playwright
pytz
fastapi
uvicorn
google-genai
python-dotenv
requests
